+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This change log contains the highlights of major changes. For details on all fin
 
 Features:
 
+- Added auxiliary `LocalTime`, `UniversalTime`, `UnixTime` class functions.
 - Published `DetectInvalidLocalTimes` property. Available at design time.
 - Added `DatabasePath` published property. Allows specifying the database path (file or directory) at design time, which will be automatically loaded at runtime.
 - Created a design and runtime package for Lazarus IDE.